rectify the error in data types

+++ avg_three_number.cpp
@@ -2,7 +2,7 @@
 
 int main() {
     // Write C++ code here
-    int a=0,b=0,c=0;
+    double a=0,b=0,c=0;
     std::cout<<"Enter the value of three number:";
     std::cin>>a;
     std::cin>>b;
